3/5 commit

+++ Module-2-Capstone/java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDAO.java
@@ -5,13 +5,13 @@ import com.techelevator.tenmo.model.Account;
 public interface AccountDAO {
 	
 	// Find the balance by account id
-	Double findBalanceByAccountId(Integer accountId);
+	Double findBalanceByAccountId(long accountId);
 	
 	// Increase balance of receiving user
-	Double increaseBalance(Integer accountId, Double amountToTransfer);
+	Double increaseBalance(long accountId, Double amountToTransfer);
 	
 	// Decrease balance of sending user
-	Double decreaseBalance(Integer accountId, Double amountToTransfer);
+	Double decreaseBalance(long accountId, Double amountToTransfer);
 	// We would put some kind of error message if the new balance is negative
 	void save(Account newAccount);
 }
